+++ TombEngine/Objects/TR1/Trap/Hammer.h
